Split AST module into smaller files.
The single ts-ast file had gotten pretty enormous, so split it into a
few smaller pieces to help make sense of things.
