+++ src/ts-ast/index.ts
@@ -1,0 +1,20 @@
+/**
+ * @license
+ * Copyright (c) 2017 The Polymer Project Authors. All rights reserved.
+ * This code may only be used under the BSD style license found at
+ * http://polymer.github.io/LICENSE.txt The complete set of authors may be found
+ * at http://polymer.github.io/AUTHORS.txt The complete set of contributors may
+ * be found at http://polymer.github.io/CONTRIBUTORS.txt Code distributed by
+ * Google as part of the polymer project is also subject to an additional IP
+ * rights grant found at http://polymer.github.io/PATENTS.txt
+ */
+
+export * from './document';
+export * from './types';
+export * from './declarations';
+
+import {Document} from './document';
+import {Type} from './types';
+import {Declaration, Method, Property} from './declarations';
+
+export type Node = Document|Declaration|Type|Method|Property;
